Refactor Canvas into its own command
